price_optimization_tools/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.5.0
coreapi==2.3.3
coreschema==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
django==3.2.7
gunicorn==21.0.0
idna==3.4
psycopg2-binary==2.9.8
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
urllib3==2.0.4
